# backend/api/docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Foodgram'
  version: ''
paths:
...
  /api/recipes/{id}/:
    get:
      operationId: Получение рецепта
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: ''
      tags:
        - Рецепты
    patch:
      operationId: Обновление рецепта
      security:
        - Token: [ ]
      description: 'Доступно только автору данного рецепта'
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта."
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: 'Рецепт успешно обновлен'
        '400':
          $ref: '#/components/responses/NestedValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
    delete:
      operationId: Удаление рецепта

      description: 'Доступно только автору данного рецепта'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '204':
          description: 'Рецепт успешно удален'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
  /api/recipes/{id}/get-link/:
    get:
      operationId: Получить короткую ссылку на рецепт
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор рецепта."
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeGetShortLink'
          description: 'Ссылка на рецепт'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
  /api/recipes/{id}/favorite/:
    post:
      operationId: Добавить рецепт в избранное
      description: 'Доступно только авторизованному пользователю.'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeMinified'
          description: 'Рецепт успешно добавлен в избранное'
        '400':
          description: 'Ошибка добавления в избранное (Например, когда рецепт уже есть в избранном)'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/RecipeNotFound'

      tags:
        - Избранное
    delete:
      operationId: Удалить рецепт из избранного
      description: 'Доступно только авторизованным пользователям'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта."
          schema:
            type: string
      responses:
        '204':
          description: 'Рецепт успешно удален из избранного'
        '400':
          description: 'Ошибка удаления из избранного (Например, когда рецепта там не было)'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/RecipeNotFound'
      tags:
        - Избранное
...
  /api/users/me/avatar/:
    put:
      operationId: Добавление аватара
      description: Добавление аватара текущего пользователя
      parameters: []
      security:
        - Token: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAvatar'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetAvatarResponse'
          description: 'Аватар успешно добавлен'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Пользователи
    delete:
      operationId: Удаление аватара
      description: Удаление аватара текущего пользователя
      parameters: []
      security:
        - Token: []
      responses:
        '204':
          description: Аватар успешно удален
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Пользователи
...
  /api/ingredients/:
    get:
      operationId: Список ингредиентов
      description: 'Список ингредиентов с возможностью поиска по имени.'
      parameters:
        - name: name
          required: false
          in: query
          description: Поиск по частичному вхождению в начале названия ингредиента.
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
          description: ''
      tags:
        - Ингредиенты
...
  /api/auth/token/logout/:
    post:
      operationId: Удаление токена
      description: Удаляет токен текущего пользователя
      parameters: []
      security:
        - Token: []
      requestBody:
        content:
          application/json:
            schema: {}

      responses:
        '204':
          content:
            application/json:
              schema: {}
          description: ''
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Пользователи
components:
  schemas:
    User:
      description:  'Пользователь (В рецепте - автор рецепта)'
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: "Адрес электронной почты"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Уникальный юзернейм"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
          description: "Имя"
          example: "Вася"
        last_name:
          type: string
          maxLength: 150
          description: "Фамилия"
          example: "Иванов"
        is_subscribed:
          type: boolean
          readOnly: true
          description: "Подписан ли текущий пользователь на этого"
          example: false
        avatar:
          type: string
          format: uri
          description: 'Ссылка на аватар'
          example: 'http://foodgram.example.org/media/users/image.png'
      required:
        - username
...
    SetAvatar:
      description: 'Добавление аватара пользователя'
      type: object
      properties:
        avatar:
          description: 'Картинка, закодированная в Base64'
          example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABAgMAAABieywaAAAACVBMVEUAAAD///9fX1/S0ecCAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAACklEQVQImWNoAAAAggCByxOyYQAAAABJRU5ErkJggg=='
          type: string
          format: binary
      required:
        - avatar
    SetAvatarResponse:
      type: object
      properties:
        avatar:
          type: string
          format: uri
          description: 'Ссылка на аватар'
          example: 'http://foodgram.example.org/media/users/image.png'
    Tag:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 32
          description: 'Уникальное название'
          example: 'Завтрак'
        slug:
          type: string
          nullable: true
          maxLength: 32
          pattern: ^[-a-zA-Z0-9_]+$
          description: 'Уникальный слаг'
          example: 'breakfast'
    RecipeList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Уникальный id'
        tags:
          description: 'Список тегов'
          readOnly: true
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        author:
          readOnly: true
          $ref: '#/components/schemas/User'
        ingredients:
          readOnly: true
          description: 'Список ингредиентов'
          type: array
          items:
            $ref: '#/components/schemas/IngredientInRecipe'
        is_favorited:
          readOnly: true
          type: boolean
          description: 'Находится ли в избранном'
        is_in_shopping_cart:
          readOnly: true
          type: boolean
          description: 'Находится ли в корзине'
        name:
          readOnly: true
          type: string
          maxLength: 256
          description: 'Название'
        image:
          readOnly: true
          description: 'Ссылка на картинку на сайте'
          example: 'http://foodgram.example.org/media/recipes/images/image.png'
          type: string
          format: uri
        text:
          readOnly: true
          description: 'Описание'
          type: string
        cooking_time:
          readOnly: true
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
    RecipeMinified:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Уникальный id'
        name:
          type: string
          maxLength: 256
          description: 'Название'
        image:
          description: 'Ссылка на картинку на сайте'
          example: 'http://foodgram.example.org/media/recipes/images/image.png'
          type: string
          format: uri
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
    RecipeGetShortLink:
      type: object
      properties:
        short-link:
          type: string
          description: 'Сокращенная ссылка'
          format: uri
          example: 'https://foodgram.example.org/s/3d0'
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 128
          example: 'Капуста'
        measurement_unit:
          type: string
          maxLength: 64
          example: 'кг'
      required:
        - name
        - measurement_unit
    IngredientInRecipe:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 128
          description: 'Название'
          example: 'Картофель отварной'
        measurement_unit:
          type: string
          maxLength: 64
          description: 'Единицы измерения'
          example: 'г'
        amount:
          type: integer
          description: 'Количество'
          minimum: 1

      required:
        - name
        - measurement_unit
...
    RecipeUpdate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        ingredients:
          description: Список ингредиентов
          type: array
          items:
            example:
              id: 1123
              amount: 10
            type: object
            properties:
              id:
                description: 'Уникальный id'
                type: integer
              amount:
                description: 'Количество в рецепте'
                type: integer
            required:
              - id
              - amount
        tags:
          description: 'Список id тегов'
          type: array
          example: [1, 2]
          items:
            type: integer
        image:
          description: 'Картинка, закодированная в Base64'
          example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABAgMAAABieywaAAAACVBMVEUAAAD///9fX1/S0ecCAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAACklEQVQImWNoAAAAggCByxOyYQAAAABJRU5ErkJggg=='
          type: string
          format: binary
        name:
          description: 'Название'
          type: string
          maxLength: 256
        text:
          description: 'Описание'
          type: string
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
      required:
        - ingredients
        - tags
        - name
        - text
        - cooking_time
    ValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        field_name:
          description: 'Название поля, в котором произошли ошибки. Таких полей может быть несколько'
          example: [ 'Обязательное поле.' ]
          type: array
          items:
            type: string
    NestedValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        ingredients:
          description: 'Ошибки в ингредиентах. В приведенном примере в первом и третьем ингредиенте не было ошибок (amount >= 1), а во втором были.'
          example: [{},{"amount":["Убедитесь, что это значение больше либо равно 1."]}, {}]
          type: array
          items:
            type: object
            properties:
              amount:
                type: array
                items:
                  type: string
    AuthenticationError:
      description: Пользователь не авторизован
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
    PermissionDenied:
      description: Недостаточно прав
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "У вас недостаточно прав для выполнения данного действия."
          type: string
    NotFound:
      description: Объект не найден
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Страница не найдена."
          type: string
  responses:
    ValidationError:
      description: 'Ошибки валидации в стандартном формате DRF'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    NestedValidationError:
      description: 'Ошибки валидации в стандартном формате DRF, в том числе с внутренними элементами.'
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/NestedValidationError'
              - $ref: '#/components/schemas/ValidationError'
    AuthenticationError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
    PermissionDenied:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
    RecipeNotFound:
      description: 'Рецепт не найден'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token